Number precision fixes for decoder

+++ LIN_Database_Encode_Decode_Tool/ldf_parser_dependencies/signal_encoding_type.hpp
@@ -25,12 +25,12 @@ public:
 	// Getters
 	std::string getName() const { return name; }
 	std::string getUnitFromRawValue(int64_t rawValue) const { return std::get<1>(getTypeInfoFromRawValue(rawValue)); }
-	int getOffsetFromPhysicalValue(double physicalValue) const { return std::get<0>(getTypeInfoFromPhysicalValue(physicalValue))[3]; }
-	int getFactorFromPhysicalValue(double physicalValue) const { return std::get<0>(getTypeInfoFromPhysicalValue(physicalValue))[2]; }
+    double getOffsetFromPhysicalValue(double physicalValue) const { return std::get<0>(getTypeInfoFromPhysicalValue(physicalValue))[3]; }
+    double getFactorFromPhysicalValue(double physicalValue) const { return std::get<0>(getTypeInfoFromPhysicalValue(physicalValue))[2]; }
 	int getMaxValueFromPhysicalValue(double physicalValue) const { return std::get<0>(getTypeInfoFromPhysicalValue(physicalValue))[1]; }
 	int getMinValueFromPhysicalValue(double physicalValue) const { return std::get<0>(getTypeInfoFromPhysicalValue(physicalValue))[0]; }
-	int getOffsetFromRawValue(int64_t rawValue) const { return std::get<0>(getTypeInfoFromRawValue(rawValue))[3]; }
-	int getFactorFromRawValue(int64_t rawValue) const { return std::get<0>(getTypeInfoFromRawValue(rawValue))[2]; }
+	double getOffsetFromRawValue(int64_t rawValue) const { return std::get<0>(getTypeInfoFromRawValue(rawValue))[3]; }
+	double getFactorFromRawValue(int64_t rawValue) const { return std::get<0>(getTypeInfoFromRawValue(rawValue))[2]; }
 	int getMaxValueFromRawValue(int64_t rawValue) const { return std::get<0>(getTypeInfoFromRawValue(rawValue))[1]; }
 	int getMinValueFromRawValue(int64_t rawValue) const { return std::get<0>(getTypeInfoFromRawValue(rawValue))[0]; }
 	ValueType getValueTypeFromRawValue(int64_t rawValue) const { return std::get<2>(getTypeInfoFromRawValue(rawValue)); }
@@ -44,9 +44,9 @@ public:
 private:
 
 	std::string name;
-	std::vector<std::tuple<std::array<int, 4>, std::string, ValueType> > encodingTypes;
-	std::tuple<std::array<int, 4>, std::string, ValueType> getTypeInfoFromRawValue(int64_t& rawValue) const;
-	std::tuple<std::array<int, 4>, std::string, ValueType> getTypeInfoFromPhysicalValue(double& physicalValue) const;
+	std::vector<std::tuple<std::array<double, 4>, std::string, ValueType> > encodingTypes;
+	std::tuple<std::array<double, 4>, std::string, ValueType> getTypeInfoFromRawValue(int64_t& rawValue) const;
+	std::tuple<std::array<double, 4>, std::string, ValueType> getTypeInfoFromPhysicalValue(double& physicalValue) const;
 
 };
